# public/openapi.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - url: https://caterify.xyz/
info:
  version: "1.0.0"
  title: Caterify API Documentation
  description: Sesuai namanya, ini docs buat API Caterify
paths:
...
  '/api/public/caterings/':
    get:
      tags:
        - Catering
      description: Returns all caterings
      summary: Get all caterings
      parameters:
        - name: api_key
          in: query
          description: API key provided by Farrel Anshary
          required: true
          schema:
            type: string
          example: asdasd
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Love letter from server ❤️
                    example: "Successfully get all caterings"
                  data:
                    type: object
                    description: Returned data from server
                    properties:
                      caterings:
                        type: array
                        description: Single catering data
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              description: ID represents a unique identifier of the catering
                              example: 11
                            name:
                              type: string
                              description: Name of the catering
                              example: "Joanda's Catering"
                            email:
                              type: string
                              description: Email of the catering
                              example: "joandacatering@gmail.com"
                            phone:
                              type: string
                              description: Phone number of the catering
                              example: "08112233"
                            image:
                              type: string
                              description: Image of the catering (nullable)
                              example: "59240f5242fa9312b4caf7a8d59f8ef7.png"
                            address:
                              type: string
                              description: Address of the catering
                              example: "Green Office Park 9, BSD"
                            radius:
                              type: number
                              format: double
                              description: Radius of supported delivery by the catering
                              example: 10000
                            latitude:
                              type: number
                              format: double
                              description: Latitude of the catering
                              example: -6.269945
                            longitude:
                              type: number
                              format: double
                              description: Longitude of the catering
                              example: 106.653181
                            average_price:
                              type: number
                              format: double
                              description: Distance of the catering from user
                              example: 32090.9090
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_401'
...
  '/api/public/schedules/catering/{cateringId}':
    get:
      tags:
        - Schedule
      description: Returns schedule of a catering in a day
      summary: Get schedule of a catering
      parameters:
        - name: cateringId
          in: path
          required: true
          description: Catering id acquired from caterings data
          schema:
            type: integer
          example: 32
        - name: api_key
          in: query
          description: API key provided by Farrel Anshary
          required: true
          schema:
            type: string
          example: asdasd
        - name: date
          in: query
          description: Date of
          required: true
          schema:
            type: string
            format: date
          example: "2021-10-31"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Love letter from server ❤️
                    example: "Successfully get schedule of Joanda's Catering on 2021-11-20"
                  data:
                    type: object
                    description: Returned data from server
                    properties:
                      schedule:
                        type: object
                        description: Single schedule data
                        properties:
                          id:
                            type: integer
                            description: ID represents a unique identifier of the schedule
                            example: 11
                          date:
                            type: string
                            format: date
                            description: Date of the schedule
                            example: "2021-11-20"
                          price:
                            type: number
                            description: Price of meal on that day
                            example: 29000
                          menu:
                            type: object
                            description: Menu on that day
                            properties:
                              id:
                                type: integer
                                description: ID represents a unique identifier of the menu
                                example: 11
                              name:
                                type: string
                                description: Name of the menu
                                example: "Paket Nasi Goreng A"
                              image:
                                type: string
                                description: Image of the menu (nullable)
                                example: "59240f5242fa9312b4caf7a8d59f8ef7.png"
                              description:
                                type: string
                                description: Description of the menu
                                example: "Ini enak lho..."

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_401'
...
components:
  schemas:
    inline_response_401:
      type: object
      properties:
        message:
          type: string
          description: Love letter from server ❤️
          example: Unauthorized